move nfc logic to separate class, read tag with associated domain
